Create slot listener before we fill static slots
