+++ src/lib/dfp/prepare-googletag.ts
@@ -27,6 +27,7 @@ import { init as scroll } from '../messenger/scroll';
 import { init as type } from '../messenger/type';
 import { init as viewport } from '../messenger/viewport';
 import { removeSlots } from '../remove-slots';
+import { createSlotFillListener } from './fill-slot-listener';
 import { fillStaticAdvertSlots } from './fill-static-advert-slots';
 import { onSlotLoad } from './on-slot-load';
 import { onSlotRender } from './on-slot-render';
@@ -137,6 +138,7 @@ export const init = (): Promise<void> => {
 						setPageTargeting(consentState, isSignedIn);
 					},
 					() => {
+						createSlotFillListener();
 						void fillStaticAdvertSlots();
 					},
 				);
